+++ packages/tests/all/jab Waiter global timeout err 1.ja.ts
@@ -1,0 +1,10 @@
+import { TestProvision } from "^jarun";
+
+import { getWaiter } from "../_fixture";
+import { setGlobalHardTimeout_experimental } from "^state-waiter";
+
+// wrong timeout
+
+export default async (prov: TestProvision) => {
+  setGlobalHardTimeout_experimental(true as any);
+};