+++ packages/state-waiter/CHANGELOG.md
@@ -11,5 +11,8 @@
   - Previously it was only possible to change timeout in calls to `await`.
 - Add the timeout amount to timeout-error-messages.
 - export `WaiterDeps`
+- Add setGlobalHardTimeout_experimental. It can set the default timeout in
+  `Waiter`, which is otherwise impossible when `Waiter` is composed by other
+  objects.
 
 ## :boom: Breaking changes